Remove dependency on bash_completion
If available, bash_completion's improvements for handling arguments
containing colons will be used, but completion will still be functional
without them.

Also unset the 'hostcomplete' shell option, as it makes it impossible
to provide completions for arguments containing an @-sign. If
bash_completion is in use, this option will already have been disabled
and replaced with something better.